backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.11.1
pytest==7.4.0

# Seguridad y Autenticación
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
argon2-cffi==23.1.0

# 2FA
pyotp==2.9.0
qrcode[pil]==7.4.2

# Validationes
email-validator>=2.1.0